backend/requirements.txt
fastapi
uvicorn
opencv-python
moviepy
librosa
torch
torchaudio
transformers
openai-whisper
python-multipart
requests
